[Decompilation] REYHI*.DAT: Loading only the high score
Still finding more obscure Turbo C++ code generation details.

Part of P0213, funded by Ember2528 and GhostRiderCog.

+++ th01/main_19.cpp
@@ -43,7 +43,7 @@ extern char rank;
 
 // Returns the high score for the difficulty previously loaded by
 // scoredat_load().
-uint32_t scoredat_hiscore_get()
+long scoredat_hiscore_get()
 {
 	return scoredat_score[0];
 }
@@ -118,3 +118,4 @@ void pascal near str_from_swapped_kanji(
 }
 
 #include "th01/hiscore/regist.cpp"
+#include "th01/main/hiscore.cpp"
